Refactor commands to separate modules
