+++ lib/args.js
@@ -2,6 +2,8 @@ var path = require('path'),
   yargs = require('yargs'),
   loadConfig = require('../lib/load_config.js');
 
+var commands = require('./commands');
+
 function commonOptions(parser) {
   return parser.option('shallow', {
     describe: 'shallow mode turns off dependency resolution, ' +
@@ -60,18 +62,22 @@ function outputOptions(parser) {
   });
 }
 
+function addCommands(parser) {
+  for (var cmd in commands) {
+    parser = parser.command(cmd, commands[cmd].description)
+  }
+  return parser
+}
+
 function parse(args) {
   // reset() needs to be called at parse time because the yargs module uses an
   // internal global variable to hold option state
-  var command = yargs.reset()
+  var command = addCommands(yargs.reset()
     .usage('Usage: $0 <command> [options]')
     .demand(1)
-    .command('build', 'build documentation')
-    .command('lint', 'check for common style and uniformity mistakes')
-    .command('serve', 'generate, update, and display HTML documentation')
     .version(function () {
       return require('../package').version;
-    })
+    }))
     .parse(args)._[0];
 
   if (command === 'build') {
