+++ lib/commands/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  'build': require('./build'),
+  'serve': require('./serve'),
+  'lint': require('./lint')
+}
+
